using eslint v9 now, updated config accordingly
