+++ tsconfig.json
@@ -7,6 +7,7 @@
     "strictNullChecks": true,
     "strictFunctionTypes": true,
     "noImplicitAny": true,
+    "resolveJsonModule": true
   },
   "exclude": [
     "node_modules",
